:recycle: split messages, take out format command
