+++ src/main/kotlin/com/github/pool_party/pull_party_bot/commands/messages/SuccessMessages.kt
@@ -1,0 +1,26 @@
+package com.github.pool_party.pull_party_bot.commands.messages
+
+val ON_LIST_SUCCESS =
+    """
+    The parties I know:
+
+
+    """.trimIndent()
+
+val ON_ARGUMENT_LIST_SUCCESS =
+    """
+    Matched parties:
+
+
+    """.trimIndent()
+
+val ON_CLEAR_SUCCESS =
+    """
+    Life is a party. Chat is dead then. 😭
+
+    All parties are over. Hope you did it only to rock even bigger ones
+    Let's rest just a little and I'll be waiting for your new entries
+    Use /create command to start all over again
+
+    Type `/help create` for more information
+    """.trimIndent()
