+++ src/main/kotlin/com/github/pool_party/pull_party_bot/commands/messages/Template.kt
@@ -1,0 +1,11 @@
+package com.github.pool_party.pull_party_bot.commands.messages
+
+// Comment template:
+// """
+// Interaction message
+// *Blank line*
+// Try to predict possible action led to this mistake and make a suggestion
+// Small command usage guide
+// *Blank line*
+// /help command suggestion
+// """
